packages/limen-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "limen"
version = "0.1.0"
requires-python = ">=3.11"
license = "MIT"
authors = [{name = "Limen contributors"}]
description = "x402 paywall for AI agent traffic. Python SDK + proxy for FastAPI, Flask, Django, Starlette."
readme = "README.md"
keywords = ["x402", "paywall", "usdc", "base", "solana", "ai-agents", "micropayments"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP",
    "Operating System :: OS Independent",
    "Typing :: Typed",
]
dependencies = [
    "pydantic[email]>=2.8",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "typer>=0.12",
    "uvicorn[standard]>=0.30",
    "redis>=5",
    "web3>=6",
    "eth-account>=0.11",
    "solders>=0.20",
    "solana>=0.34",
    "prometheus-client>=0.20",
    "structlog>=24.1",
    "tenacity>=8.3",
    "ulid-py>=1.1",
    "opentelemetry-api>=1.25",
]

[project.urls]
Homepage = "https://limen.dev"
Documentation = "https://github.com/dhruvalgupta2003/limen/tree/main/docs"
Repository = "https://github.com/dhruvalgupta2003/limen"

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5",
    "respx>=0.21",
    "ruff>=0.6",
    "mypy>=1.11",
    "pynacl>=1.5",
]

[project.scripts]
limen = "limen.cli:main"

[project.optional-dependencies]
fastapi = ["fastapi>=0.110"]
flask = ["flask>=3.0"]
django = ["django>=5.0"]
starlette = ["starlette>=0.37"]
all = [
    "fastapi>=0.110",
    "flask>=3.0",
    "django>=5.0",
    "starlette>=0.37",
]

[tool.hatch.build.targets.wheel]
packages = ["limen"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "RUF",  # ruff
]
ignore = [
    "E501",  # line too long — line-length already set above
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011", "N802"]

[tool.mypy]
strict = true
python_version = "3.11"
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
disallow_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = [
    "solders.*",
    "solana.*",
    "web3.*",
    "eth_account.*",
    "ulid.*",
    "redis.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-ra --strict-markers --cov=limen --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["limen"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "@overload",
]
